--- apps/client/src/shared/types/favorite-artists-response.ts
@@ -1,9 +1,0 @@
-interface Artists {
-  artistId: string;
-  name: string;
-  profileUrl: string;
-}
-
-export interface FavoriteArtistsResponses {
-  artists: Artists[];
-}
